Untie batter animation from FPS

+++ include/Batter/Batter.hpp
@@ -9,7 +9,7 @@ private:
     int frame = 0;          // Current drawing number
     int frame_count = 5;    // Number of drawings
     int frames_to_wait = 2; // Wait this many engine frames to display the next drawing
-    int wait = 0;           // Current wait
+    float wait = 0;           // Current wait
     bool playing = false;
     void Anim_Reset();
